requirements.txt
# 核心框架
fastapi==0.115.5
uvicorn[standard]==0.34.0
pydantic==2.10.3
pydantic-settings==2.6.1

# 机器学习
xgboost==2.1.3
lightgbm==4.5.0
scikit-learn==1.8.0
imbalanced-learn==0.12.4

# 数据处理
polars==1.36.1
numpy==2.2.1
scipy==1.15.1

# 图数据库
neo4j==5.27.0

# LLM 相关 (简化版)
openai==1.59.6
langchain==0.3.13
langchain-community==0.3.13
langgraph==0.2.56

# 数据库
sqlalchemy==2.0.36
redis==5.2.1

# 工具
python-dotenv==1.0.1
pyyaml==6.0.2
python-multipart==0.0.20
httpx==0.28.1

# 监控与日志
prometheus-client==0.21.1
loguru==0.7.3

# 开发工具
pytest==8.3.4
pytest-asyncio==0.25.2
black==24.10.0
